Cythonize SimpleLabeledAssembler
