+++ khmer/_oxli/assembly.pxd
@@ -5,7 +5,8 @@ from libc.stdint cimport uint16_t
 
 from oxli_types cimport *
 from hashing cimport CpKmer, Kmer
-from graphs cimport CpHashgraph, CpHashtable, CpLabelHash, get_hashgraph_ptr
+from graphs cimport (CpHashgraph, CpHashtable, CpLabelHash, 
+                     get_hashgraph_ptr, get_labelhash_ptr)
 
 
 cdef extern from "oxli/assembler.hh" namespace "oxli":
@@ -20,7 +21,7 @@ cdef extern from "oxli/assembler.hh" namespace "oxli":
         string assemble_left(const CpKmer) const     
         string assemble_right(const CpKmer) const
 
-    cdef cppclass CpSimpleLabeledAsssembler "oxli::SimpleLabeledAsssembler":
+    cdef cppclass CpSimpleLabeledAssembler "oxli::SimpleLabeledAssembler":
         CpSimpleLabeledAssembler(const CpLabelHash *)
 
         vector[string] assemble(const CpKmer)
@@ -50,3 +51,27 @@ cdef class LinearAssembler:
     cdef str _assemble_right(self, Kmer start)
 
 
+cdef class SimpleLabeledAssembler:
+    cdef unique_ptr[CpSimpleLabeledAssembler] _this
+
+    cdef public object labels
+    cdef CpLabelHash * _label_ptr
+
+    cdef public object stop_filter
+    cdef CpHashgraph * _stop_filter_ptr
+    
+    cdef vector[string] _assemble(self, Kmer start)
+
+'''
+cdef class JunctionCountAssembler:
+    cdef unique_ptr[CpJunctionCountAssembler] _this
+
+    cdef public object graph
+    cdef CpHashgraph * _graph_ptr
+
+    cdef public object stop_filter
+    cdef CpHashgraph * _stop_filter_ptr
+    
+    cdef str _assemble(self, Kmer)
+    cdef uint16_t _consume(self, string)
+'''
