+++ khmer/_oxli/graphs.pyx
@@ -1,8 +1,7 @@
 # cython: c_string_type=unicode, c_string_encoding=utf8
 from cython.operator cimport dereference as deref
 
-from .._khmer import Countgraph
-from .._khmer import Nodegraph
+from .._khmer import Countgraph, Nodegraph, GraphLabels 
 
 
 cdef CpHashgraph * get_hashgraph_ptr(object graph):
@@ -12,3 +11,10 @@ cdef CpHashgraph * get_hashgraph_ptr(object graph):
     cdef CPyHashgraph_Object* ptr = <CPyHashgraph_Object*> graph
     return deref(ptr).hashgraph
     
+
+cdef CpLabelHash * get_labelhash_ptr(object labels):
+    if not isinstance(labels, GraphLabels):
+        return NULL
+    
+    cdef CPyGraphLabels_Object * ptr = <CPyGraphLabels_Object*> labels
+    return deref(ptr).labelhash
