+++ src/khmer/_cpy_assembly.cc
@@ -7,147 +7,6 @@ using namespace oxli;
 
 namespace khmer {
 
-PyTypeObject khmer_KSimpleLabeledAssembler_Type = {
-    PyVarObject_HEAD_INIT(NULL, 0)  /* init & ob_size */
-    "_khmer.SimpleLabeledAssembler",            /* tp_name */
-    sizeof(khmer_KSimpleLabeledAssembler_Object), /* tp_basicsize */
-    0,                       /* tp_itemsize */
-    (destructor)khmer_simplelabeledassembler_dealloc, /* tp_dealloc */
-    0,                       /* tp_print */
-    0,                       /* tp_getattr */
-    0,                       /* tp_setattr */
-    0,                       /* tp_compare */
-    0,                       /* tp_repr */
-    0,                       /* tp_as_number */
-    0,                       /* tp_as_sequence */
-    0,                       /* tp_as_mapping */
-    0,                       /* tp_hash */
-    0,                       /* tp_call */
-    0,                       /* tp_str */
-    0,                       /* tp_getattro */
-    0,                       /* tp_setattro */
-    0,                       /* tp_as_buffer */
-    Py_TPFLAGS_DEFAULT | Py_TPFLAGS_BASETYPE,   /* tp_flags */
-    0,                       /* tp_doc */
-    0,                       /* tp_traverse */
-    0,                       /* tp_clear */
-    0,                       /* tp_richcompare */
-    0,                       /* tp_weaklistoffset */
-    0,                       /* tp_iter */
-    0,                       /* tp_iternext */
-    khmer_simplelabeledassembler_methods, /* tp_methods */
-    0,                       /* tp_members */
-    0,                       /* tp_getset */
-    0,                       /* tp_base */
-    0,                       /* tp_dict */
-    0,                       /* tp_descr_get */
-    0,                       /* tp_descr_set */
-    0,                       /* tp_dictoffset */
-    0,                       /* tp_init */
-    0,                       /* tp_alloc */
-    khmer_simplelabeledassembler_new,      /* tp_new */
-};
-
-PyMethodDef khmer_simplelabeledassembler_methods[] = {
-    {
-        "assemble",
-        (PyCFunction)simplelabeledassembler_assemble, METH_VARARGS | METH_KEYWORDS,
-        "Assemble paths, using labels to jump branches."
-    },
-    {NULL, NULL, 0, NULL}           /* sentinel */
-};
-
-
-void khmer_simplelabeledassembler_dealloc(khmer_KSimpleLabeledAssembler_Object *
-        obj)
-{
-    delete obj->assembler;
-    obj->assembler = NULL;
-
-    Py_TYPE(obj)->tp_free((PyObject*)obj);
-}
-
-
-PyObject * khmer_simplelabeledassembler_new(PyTypeObject *type,
-        PyObject *args,
-        PyObject *kwds)
-{
-    khmer_KSimpleLabeledAssembler_Object *self;
-    self = (khmer_KSimpleLabeledAssembler_Object*)type->tp_alloc(type, 0);
-
-    if (self != NULL) {
-        PyObject * labelhash_o;
-        LabelHash * labelhash = NULL;
-
-        if (!PyArg_ParseTuple(args, "O", &labelhash_o)) {
-            Py_DECREF(self);
-            return NULL;
-        }
-
-        if (PyObject_TypeCheck(labelhash_o, &khmer_KGraphLabels_Type)) {
-            khmer_KGraphLabels_Object * klo = (khmer_KGraphLabels_Object *) labelhash_o;
-            labelhash = klo->labelhash;
-        } else {
-            PyErr_SetString(PyExc_ValueError,
-                            "SimpleLabeledAssembler needs a GraphLabels object.");
-            Py_DECREF(self);
-            return NULL;
-        }
-
-        try {
-            self->assembler = new SimpleLabeledAssembler(labelhash);
-        } catch (std::bad_alloc &e) {
-            Py_DECREF(self);
-            return PyErr_NoMemory();
-        }
-
-    }
-
-    return (PyObject *) self;
-}
-
-
-
-PyObject *
-simplelabeledassembler_assemble(khmer_KSimpleLabeledAssembler_Object * me,
-                                PyObject * args, PyObject *kwargs)
-{
-    SimpleLabeledAssembler * assembler = me->assembler;
-
-    PyObject * val_o;
-    khmer_KNodegraph_Object * nodegraph_o = NULL;
-    Nodegraph * stop_bf = NULL;
-
-    const char *kwnames[] = {"seed_kmer", "stop_filter", NULL};
-
-    if (!PyArg_ParseTupleAndKeywords(args, kwargs, "O|O!",
-                                     const_cast<char **>(kwnames),
-                                     &val_o, &khmer_KNodegraph_Type,
-                                     &nodegraph_o)) {
-        return NULL;
-    }
-
-
-    Kmer start_kmer;
-    if (!ht_convert_PyObject_to_Kmer(val_o, start_kmer, assembler->graph)) {
-        return NULL;
-    }
-
-    if (nodegraph_o) {
-        stop_bf = nodegraph_o->nodegraph;
-    }
-
-    std::vector<std::string> contigs = assembler->assemble(start_kmer, stop_bf);
-
-    PyObject * ret = PyList_New(contigs.size());
-    for (unsigned int i = 0; i < contigs.size(); i++) {
-        PyList_SET_ITEM(ret, i, PyUnicode_FromString(contigs[i].c_str()));
-    }
-
-    return ret;
-}
-
-
 
 /********************************
  * JunctionCountAssembler
